[code]history changes

+++ xuberservice/src/main/java/com/gox/xuberservice/model/UpdateRequest.kt
@@ -58,7 +58,7 @@ data class UpdateRequest(
             val user_rated: Double? = 0.0
     ) {
         data class Payment(
-                val card: Int? = 0,
+                val card: Double? = 0.0,
                 val cash: Double? = 0.0,
                 val commision: Double? = 0.0,
                 val commision_percent: Double? = 0.0,
